requirements.txt
paho-mqtt
locust
pycryptodome
cryptography
redis
ansible
protobuf